remove VisualTestUtils
